+++ src/view/InputView.js
@@ -1,12 +1,12 @@
+import { InputValidator } from '../helpers/index.js';
 import { INPUT_MESSAGES } from '../lib/constants.js';
 import { readLineAsync, retryUntilSuccess } from '../lib/utils.js';
-import Validator from './Validator.js';
 
 export default class InputView {
   static async getNames() {
     return retryUntilSuccess(async () => {
       const rawName = await readLineAsync(INPUT_MESSAGES.carName);
-      Validator.validateCarName(rawName);
+      InputValidator.validateCarName(rawName);
 
       return rawName.split(',');
     });
@@ -15,7 +15,7 @@ export default class InputView {
   static async getCount() {
     return retryUntilSuccess(async () => {
       const rawCount = await readLineAsync(INPUT_MESSAGES.count);
-      Validator.validateCount(rawCount);
+      InputValidator.validateCount(rawCount);
 
       return Number(rawCount);
     });
